+++ apps/web/app/composables/index.ts
@@ -107,3 +107,4 @@ export * from './useBlockClasses';
 export * from './useCallisto';
 export * from './useCancellationForm';
 export * from './useLogEvent';
+export * from './useEditorPreview';
