+++ apps/web/app/composables/useEditorPreview/types.ts
@@ -1,0 +1,1 @@
+export type PreviewDevice = 'mobile' | 'tablet' | 'desktop';
